+++ compiler/src/arm_arch_full.ml
@@ -26,6 +26,9 @@ module Arm_core = struct
 
   let known_implicits = ["NF", "_nf_"; "ZF", "_zf_"; "CF", "_cf_"; "VF", "_vf_"]
 
+  let alloc_stack_need_extra sz =
+    not (Arm_params_common.ARMFopn.is_arith_small (Conv.cz_of_z sz))
+
 end
 
 module Arm (Lowering_params : Arm_input) : Arch_full.Core_arch = struct