+++ compiler/src/compile.ml
@@ -82,12 +82,20 @@ let compile (type reg regx xreg rflag cond asm_op extra_op)
 
     CheckAnnot.check_stack_size fds;
 
+
+    let get_internal_size _fd sfe =
+      let stk_size =
+        BinInt.Z.add sfe.Expr.sf_stk_sz sfe.Expr.sf_stk_extra_sz in
+      Conv.z_of_cz (Memory_model.round_ws sfe.sf_align stk_size)
+    in
+
     let fds =
       Regalloc.alloc_prog translate_var
         (fun _fd extra ->
           match extra.Expr.sf_save_stack with
           | Expr.SavedStackReg _ | Expr.SavedStackStk _ -> true
           | Expr.SavedStackNone -> false)
+        get_internal_size
         fds
     in
     let fds = List.map (fun (y, _, x) -> (y, x)) fds in