+++ compiler/src/x86_arch_full.ml
@@ -41,6 +41,7 @@ module X86_core = struct
 
   let known_implicits = ["OF","_of_"; "CF", "_cf_"; "SF", "_sf_"; "ZF", "_zf_"]
 
+  let alloc_stack_need_extra _ = false
 end